RLIB-23: Add more expected output for RLIB 2.0.x
The delayed="yes" XML parameter was added back to src/common-xml/graph.xml
Although this is ignored by RLIB 3.0.x, this is needed to create
identical output for the "graph" test case.

The graph HTML output is accepted with the difference that the
RLIB 2.0.x HTML output driver doesn't add the necessary <span>
tag for delayed data, which is a bug.